feat: switch to signal inputs
This commit switches inputs to signal inputs to be compatible with zoneless change detection
because everything should be a signal primitive. As such, its changes may be caught by the
execution context. Every time a signal changes, it would schedule a change detection event,
even in zoneless mode.

I didn't update `@Output()` to `output` because we still use `.observed` properties.