fix: handle empty value returned by getSemanticHTML
A regression was spotted by our team when updating to the lastest ngx-quill version, where required form controls applied on quill editors were no longer be marked invalid when emptied.

It looks like `getSemanticHTML()` can return `<p></p>` when the editor is emptied (at least when it is emptied programatically using `setText('')`).
This commit augments the "empty" check to handle this possible value and `<div></div>`.
A unit test has been added to reproduce this issue.