fix: reduce asynchronous stuff to resolve flickering
In this commit, we introduce a slightly smarter approach to handling custom module registration
and executing the `beforeRender` function. The current issue is that the `quill-editor-toolbar`
may be projected before the editor is rendered due to the `Promise.all` in `mergeMap`.
This is noticeable on slower devices, where the toolbar appears first (without styles),
and then after a microtask delay, Quill is created.

Switching to observables resolves this issue because observables can emit both synchronously and
asynchronously, and most observables can cache their emissions using `shareReplay`,
ensuring that subsequent emissions are always synchronous.