refactor: move `quill` to separate entry point (#1981)
Modules that are wrapped with a dynamic import are always marked as side-effectful. For example, if we do `import('lodash-es').then(m => m.sortBy(...))`, it will bundle **all** functions from that module, regardless of whether it's in ESM format or not. This happens because the compiler cannot detect which static imports are used from the module.

Having separate entry points helps the compiler detect static imports and tree-shake functions and classes that are not referenced. A static import like `import { ... } from 'quill'` has no side effects, and only the `quill.ts` file is marked as side-effectful.